// src/data/CardData.d.ts
export declare const content: {
    title: string;
    link: string;
    video: string;
    image: string;
    alt: string;
    date: string;
    excerpt: string;
    tags: {
        category: {
            id: number;
            name: string;
            slug: string;
        }[];
    };
};